fix(ai): split spend helpers under the r8 loc cap, tighten a one-sided guard
commands/ai/mod.rs had grown past R8's 1400-LOC hard cap after the #1161
fixes, failing the architecture test CI gates on. Split the pure/AppHandle-
free spend helpers (spend_summary_from_store, spend_summary_value,
spend_totals_json, resolve_window_days, per_provider_with_zero_rows) into
commands/ai/spend.rs, same shape as commands::match_resume's constraints
split; the #[tauri::command] itself stays in mod.rs so it's still reachable
at commands::ai::ai_spend_summary. Their tests move with them into spend.rs's
own #[cfg(test)] mod, and the call-site regression test gains a same-window
row so it can no longer pass with an always-zero today total.

Refs #1159
Refs #1161
