+++ apps/desktop/src-tauri/src/commands/ai/test.rs
@@ -341,176 +341,8 @@ fn unrelated_jobs_do_not_block_indexing() {
     }
 }
 
-// ── spend_totals_json (ai_spend_summary today/windowTotals shape, #1161) ──────
-
-use crate::spend::SpendTotals;
-
-#[test]
-fn spend_totals_json_carries_the_exact_totals_given() {
-    let totals = SpendTotals {
-        input_tokens: 12_431,
-        output_tokens: 3_204,
-        est_cost_usd: 0.42,
-    };
-
-    let out = spend_totals_json(totals);
-    assert_eq!(out["inputTokens"], 12_431);
-    assert_eq!(out["outputTokens"], 3_204);
-    assert_eq!(out["estCostUsd"], 0.42);
-}
-
-#[test]
-fn spend_summary_value_labels_today_and_window_totals_from_their_own_input() {
-    // Covers `spend_summary_value`'s payload SHAPING only (it just labels
-    // whatever two `SpendTotals` it's handed) — it does NOT exercise the
-    // call site that decides what those two values ARE. That's
-    // `ai_spend_summary_call_site_keeps_today_and_window_totals_distinct`
-    // below (issue #1161's C1-r2-RBA-2): this test alone would stay green
-    // even if the call site collapsed both to `totals_since(window_start)`.
-    let today = SpendTotals {
-        input_tokens: 100,
-        output_tokens: 50,
-        est_cost_usd: 0.10,
-    };
-    let window_totals = SpendTotals {
-        input_tokens: 9_000,
-        output_tokens: 4_000,
-        est_cost_usd: 12.0,
-    };
-
-    let out = spend_summary_value(today, window_totals, vec![], vec![], json!({}));
-    assert_eq!(out["today"]["inputTokens"], 100);
-    assert_eq!(out["windowTotals"]["inputTokens"], 9_000);
-    assert_ne!(out["today"], out["windowTotals"]);
-}
-
-#[test]
-fn ai_spend_summary_call_site_keeps_today_and_window_totals_distinct() {
-    // Regression for C1-r2-RBA-2: the tautology above only checks that
-    // `spend_summary_value` labels its two inputs correctly — it never calls
-    // the actual `ai_spend_summary` call site, so reverting mod.rs back to
-    // `today = store.totals_since(window_start)` (C1-r1-RBA-1's original
-    // defect) would leave the whole suite green. This drives the real call
-    // site (`spend_summary_from_store`) against a real on-disk `SpendStore`
-    // seeded with a row outside "today" but inside a 7-day window, so a
-    // collapse back onto one query fails here.
-    use crate::data_store::DataStore;
-    use crate::spend::SpendStore;
-
-    let dir = TempDir::new().unwrap();
-    let store = SpendStore::open(&dir.path().to_path_buf()).unwrap();
-
-    let five_days_ago = crate::db::now_ms() - 5 * 86_400_000;
-    store
-        .import(&serde_json::json!([{
-            "id": "spend-c1-r2-rba-2",
-            "createdAt": five_days_ago,
-            "provider": "openai",
-            "model": "gpt-test",
-            "inputTokens": 500,
-            "outputTokens": 200,
-            "estCostUsd": 3.0,
-        }]))
-        .unwrap();
-
-    let out = spend_summary_from_store(&store, 7);
-    assert_eq!(
-        out["today"]["inputTokens"], 0,
-        "a 5-day-old row must not count as today's spend"
-    );
-    assert_eq!(
-        out["windowTotals"]["inputTokens"], 500,
-        "the same row must count inside a 7-day window"
-    );
-    assert_ne!(out["today"], out["windowTotals"]);
-}
-
-#[test]
-fn resolve_window_days_defaults_to_one_and_clamps_to_the_max() {
-    assert_eq!(resolve_window_days(None), 1, "unset means today only");
-    assert_eq!(
-        resolve_window_days(Some(0)),
-        1,
-        "zero-day window is not valid"
-    );
-    assert_eq!(
-        resolve_window_days(Some(7)),
-        7,
-        "in-range values pass through"
-    );
-    assert_eq!(
-        resolve_window_days(Some(500)),
-        crate::spend::SPEND_WINDOW_MAX_DAYS,
-        "an oversized request must clamp, not report an unbounded window"
-    );
-    assert_eq!(
-        resolve_window_days(Some(u32::MAX)),
-        crate::spend::SPEND_WINDOW_MAX_DAYS,
-        "u32::MAX must clamp too, never resolve to an all-time window"
-    );
-}
-
-// ── per_provider_with_zero_rows (ai_spend_summary window merge, #1161) ────────
-
-use crate::spend::ProviderTotals;
-
-fn totals(provider: &str, input: u64, output: u64, cost: f64) -> ProviderTotals {
-    ProviderTotals {
-        provider: provider.to_string(),
-        input_tokens: input,
-        output_tokens: output,
-        est_cost_usd: cost,
-    }
-}
-
-#[test]
-fn a_provider_active_in_the_window_reports_its_windowed_totals_with_no_reason() {
-    let all_time = vec![totals("openai", 10_000, 5_000, 1.5)];
-    let windowed = vec![totals("openai", 10_000, 5_000, 1.5)];
-
-    let out = per_provider_with_zero_rows(all_time, windowed);
-    assert_eq!(out.len(), 1);
-    assert_eq!(out[0]["provider"], "openai");
-    assert_eq!(out[0]["inputTokens"], 10_000);
-    assert!(
-        out[0].get("reason").is_none(),
-        "an active row must not carry a reason"
-    );
-}
-
-#[test]
-fn a_provider_absent_from_the_window_becomes_a_zero_row_with_a_reason() {
-    // Historically active (all_time), but nothing in THIS window.
-    let all_time = vec![totals("anthropic", 20_000, 8_000, 2.0)];
-    let windowed = vec![]; // nothing in the window
-
-    let out = per_provider_with_zero_rows(all_time, windowed);
-    assert_eq!(out.len(), 1, "the provider must still appear");
-    assert_eq!(out[0]["provider"], "anthropic");
-    assert_eq!(out[0]["inputTokens"], 0);
-    assert_eq!(out[0]["outputTokens"], 0);
-    assert_eq!(out[0]["estCostUsd"], 0.0);
-    assert_eq!(out[0]["reason"], "no spend in window");
-}
-
-#[test]
-fn a_free_provider_absent_from_the_window_is_labelled_local() {
-    let all_time = vec![totals("ollama", 50_000, 20_000, 0.0)];
-    let out = per_provider_with_zero_rows(all_time, vec![]);
-    assert_eq!(out[0]["reason"], "local — always $0");
-}
-
-#[test]
-fn every_all_time_provider_survives_the_merge_even_with_an_empty_window() {
-    let all_time = vec![
-        totals("openai", 1, 1, 0.01),
-        totals("anthropic", 2, 2, 0.02),
-        totals("ollama", 3, 3, 0.0),
-    ];
-    let out = per_provider_with_zero_rows(all_time, vec![]);
-    assert_eq!(
-        out.len(),
-        3,
-        "no provider must be dropped by an empty window"
-    );
-}
+// AI-spend visibility (`ai_spend_summary` and its pure helpers, #1161) moved
+// to `commands::ai::spend`'s own `#[cfg(test)] mod test` alongside the code
+// it covers — see that module for `spend_totals_json`, `spend_summary_value`,
+// `spend_summary_from_store`, `resolve_window_days`, and
+// `per_provider_with_zero_rows` coverage.
